Add files via upload
A WAY TO BE A SAINT! Is there somebody who would like to be a saint in the world? If yes, you can download this application and to read about how to become a saint.
The application contains some holy people, the latin and english prayers, the Bible you can see online and my web too. There are some quotes from blessed Stefan Wyszynski (a Polish Cardinal) and 40 thoughts about Who is a modern holy person? Good luck.